10_04

+++ common/movies/views.py
@@ -2,6 +2,13 @@ from django.shortcuts import render, redirect
 from .models import Movie
 from .forms import MovieForm
 
+from django.views.decorators.http import (
+    require_http_methods, # GET & POST요청 # create, update
+    require_safe, # GET요청 # index, detail
+    require_POST, # POST요청 # delete
+)
+
+@require_safe
 def index(request):
   movies = Movie.objects.all()
   context = {
@@ -9,6 +16,7 @@ def index(request):
   }
   return render(request, 'movies/index.html', context)
 
+@require_safe
 def detail(request, pk):
   movie = Movie.objects.get(pk=pk)
   context = {
@@ -17,6 +25,7 @@ def detail(request, pk):
   return render(request, 'movies/detail.html', context)
 
 # 유효성 검사
+@require_http_methods(['GET', 'POST'])
 def create(request):
   if request.method == 'POST':
     form = MovieForm(request.POST, request.FILES)
@@ -30,6 +39,7 @@ def create(request):
   }
   return render(request, 'movies/create.html', context)
 
+@require_http_methods(['GET', 'POST'])
 def update(request, pk):
   movie = Movie.objects.get(pk=pk)
   if request.method == 'POST':
@@ -45,7 +55,10 @@ def update(request, pk):
   }
   return render(request, 'movies/update.html', context)
 
+@require_POST
 def delete(request, pk):
-  movie = Movie.objects.get(pk=pk)
-  movie.delete()
-  return redirect('movies:index')
+  if request.method == 'POST':
+    movie = Movie.objects.get(pk=pk)
+    movie.delete()
+    return redirect('movies:index')
+  return redirect('movies:detail', movie.pk)
